refactor: Menus를 일급 컬렉션으로 구현
