+++ src/main/java/kitchenpos/eatinorders/tobe/domain/menu/Menus.java
@@ -1,31 +1,34 @@
-package kitchenpos.eatinorders.tobe.domain.service;
+package kitchenpos.eatinorders.tobe.domain.menu;
 
 import kitchenpos.eatinorders.tobe.domain.Order;
 import kitchenpos.eatinorders.tobe.domain.OrderLineItem;
-import kitchenpos.eatinorders.tobe.domain.menu.Menu;
-import kitchenpos.eatinorders.tobe.domain.menu.MenuManager;
-import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Component
-public class MenuValidateService {
-    private final MenuManager menuManager;
+public class Menus {
+    private final List<Menu> menus;
 
-    public MenuValidateService(final MenuManager menuManager) {
-        this.menuManager = menuManager;
+    private Menus(final List<Menu> menus) {
+        menus.forEach(this::validateMenu);
+        this.menus = Collections.unmodifiableList(menus);
     }
 
-    void validateOrder(final Order order) {
+    public static Menus from(final MenuManager menuManager, final Order order) {
         final List<OrderLineItem> orderLineItems = order.getOrderLineItems();
         final List<Menu> menus = menuManager.getMenus(orderLineItems.stream()
                 .map(OrderLineItem::getMenuId)
                 .collect(Collectors.toList()));
-        menus.forEach(this::validateMenu);
         if (menus.size() != orderLineItems.size()) {
             throw new IllegalArgumentException("부적절한 메뉴는 주문할 수 없습니다.");
         }
+        return new Menus(menus);
+    }
+
+    public List<Menu> getMenus() {
+        return new ArrayList<>(menus);
     }
 
     private void validateMenu(final Menu menu) {
